Foi criado a conexão da aplicação com o Banco de Dados, criado também…

+++ src/src/model/Entity.java
@@ -2,13 +2,12 @@ package model;
 
 public class Entity {
 
-    private Integer id;
-
-    public int getId() {
+    private long id;
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 }
